+++ docs/source/cmd_main.rst
@@ -2,7 +2,7 @@ ya360
 -----
 
 .. argparse::
-   :module: ya360.__main__
+   :module: ya360.cmd
    :func: gen_parser
    :prog: ya360
    :nosubcommands:
