+++ docs/source/cmd_ya360_departments.rst
@@ -2,7 +2,7 @@ ya360 departments
 -----------------
 
 .. argparse::
-   :module: ya360.__main__
+   :module: ya360.cmd
    :func: gen_parser
    :prog: ya360
    :path: departments
