+++ docs/source/cmd_ya360_group.rst
@@ -2,7 +2,7 @@ ya360 group
 -----------
 
 .. argparse::
-   :module: ya360.__main__
+   :module: ya360.cmd
    :func: gen_parser
    :prog: ya360
    :path: group
