+++ docs/source/cmd_ya360_make-config.rst
@@ -2,7 +2,7 @@ ya360 make-config
 -----------------
 
 .. argparse::
-   :module: ya360.__main__
+   :module: ya360.cmd
    :func: gen_parser
    :prog: ya360
    :path: make-config
