+++ docs/source/cmd_ya360_user.rst
@@ -2,7 +2,7 @@ ya360 user
 ----------
 
 .. argparse::
-   :module: ya360.__main__
+   :module: ya360.cmd
    :func: gen_parser
    :prog: ya360
    :path: user
