+++ docs/source/usage.rst
@@ -26,9 +26,6 @@
 Конфигурирование
 ----------------
 
-.. warning::
-    Метод конфигурирования изменен!
-
 После успешной установки, утилиту необходимо настроить на работу с вашей организацией.
 При первом запуске и до момента пока не будет сформирован конфигурационный файл,
 приложение будет запускать мастер первоначальной настройки.
@@ -38,3 +35,6 @@
 .. code-block:: console
 
     $ ya360 make-config
+
+.. note::
+    Начиная с версии ``2.6.0`` приложение самостоятельно перезапрашивает токен с заканчивающимся сроком действия
