+++ ya360/__init__.py
@@ -4,5 +4,5 @@
 
 """
 
-__version__ = '2.5.0-dev0'
+__version__ = '2.5.0'
 __author__ = 'Igor Kuptsov'
